requirements.txt
asgiref             
Django              
djangorestframework 
pip                 
pytz                
setuptools          
sqlparse            